pyproject.toml
[project]
name = "sopharm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django-browser-reload>=1.21.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-storages[s3]>=1.14.6",
    "django-webpack-loader>=3.2.2",
    "django-widget-tweaks>=1.5.0",
    "wagtail>=7.2.1",
    "whitenoise>=6.11.0",
]
